requirements.txt
langchain
langchain-community
langchain-ibm
chromadb
pypdf
gradio
sentence-transformers
ibm-watsonx-ai
python-dotenv
